+++ packages/api/src/api/index.ts
@@ -4,11 +4,12 @@ import { auth } from './auth';
 import { health } from './health';
 import { users } from './users';
 import { videos } from './videos';
-import { repository } from './repository';
+import { project } from './project';
+
 export const api = Router();
 
 api.use('/health', health);
 api.use('/users', users);
 api.use('/videos', videos);
 api.use('/auth', auth);
-api.use('/repository', repository);
+api.use('/project', project);
